Back out "progress: backout D65769371 and D65138587"
Summary:
Originally D65138587 was backed out because when using a release version on Windows (as opposed as running it on dev mode), running new commands would spawn a new *visible* terminal Window on Windows. D65769371 also had to be backed out because it depended on D65138587.

Because the issue has been fixed already in D66483471, we can back out the back-outs.

Reviewed By: muirdm

Differential Revision: D66483917

fbshipit-source-id: 32a56cf3f304e80eaae250fc6810f24231f1525f