Plumb prefix/searchRoot through to edenAPI globFiles
Summary:
# Context
Returning only results that match a prefix is a feature supported by the suffix_query api. Previously we were not planning to support this feature, but now allow it for limited prefixes.

This allows us to greatly increase the speed of running the suffix query api

# This Diff
Plumb prefix through the suffix_query call path. Suffix query is equal to the searchRoot.

# Technical Details
Since we've moved www into fbsource, this query hits a wider range of intended results than when it was run inside www. Returning all these results is slow, so by cutting them out on the server side, we can greatly increase the overall speed

On the rust side, it's an optional<Vec<String>>, but since the cxx ffi doesn't support passing optionals through, we use an empty list and check on the rust side.

Reviewed By: muirdm

Differential Revision: D66501881

fbshipit-source-id: 7c87f9233aeca9d861c045769c9f4e7a8eaec145