website: upgrade to Docusaurus 3
Summary:
Docusaurus 3 has some breaking changes. Namely the "raw" HTML is no longer
supported. We have a plugin to use raw HTML to render generated colorful
output, although the coloring feature is not actually used. I disabled the
"coloring" feature to fix the MDX 3 compatibility for now. We can build our own
ANSI-to-MDX logic in the future if we want.

Reviewed By: sggutier

Differential Revision: D66508226

fbshipit-source-id: 15d5f8955924b7eb96970d9d3abc01058f2452d9