rebase: validate repo state when transitioning to on-disk mode
Summary:
In-memory rebase allows for a dirty working copy (since it doesn't use the working copy). However, when transitioning from in-memory rebase to on-disk rebase due to a conflict, we weren't validating that the working copy was clean. This lead to confusing/incorrect results when rebasing in a working copy with multiple parents. In particular, it reported the destination commit already had all the source commit's changes.

Fix by adding a `bailifchanged` when transitioning from in-memory to on-disk. I also tightened up the check around the "resuming interrupted rebase" to only ignore multiple parents in the `--continue` case.

Reviewed By: zzl0

Differential Revision: D66476017

fbshipit-source-id: d84c1dee3e513dc4309c569b3cefc4a8aad28f2c