megarepo_api: don't derive filenodes while building megarepo commits
Summary:
The filenodes derived data type only applies to public commits.  Ordinarily it is a no-op when attempted on a draft commit, however derivation batching can fail if it discovers that the parent of a batch root is not derived.

Don't even try to derive filenodes for these commits.  The derivation will happen for real when the bookmark is moved to the head commit and the ancestors all become public.

Reviewed By: gustavoavena

Differential Revision: D66507350

fbshipit-source-id: 6c4717469949aee709e7ec0e939596b2596699c1