Add mount args to validate checkouts for vscode provided workspace folders
Summary:
We are extending the health-report command to accept mount paths as arguments in order so that can limit health-report to only checking the health of the passed in mounts (i.e. eden health-report --mounts ~/fbsource ~/opsfiles ~/configerator).

In this case vscode will provide th list of worspace folder paths that are checked out by the user  and run health-report only for those set of workpsace folders.

Reviewed By: jdelliot

Differential Revision: D66507034

fbshipit-source-id: 5abc637ce6a339df8da5a15fb03b67275c9a2038