scuba_args: add --scuba-log-file argument
Summary:
Currently we can override scuba logging to a file by setting the `--scuba-dataset` parameter to `file://<FILENAME>`.

This is not ideal.  While this looks a bit like a URL, it's not a URL (and is not even a validly formed URL in some cases), which can be confusing.  Better to have a separate command line flag for setting the scuba output to be a log file.

Rework the scuba logging args as follows:

* Each binary can optionally define a default scuba dataset.
* The command line arguments may specify `--scuba-dataset DATASET` to divert to a different scuba table.
* The command line arguments may specify `--scuba-log-file FILENAME` to log to a file.
* The command line arguments may specify `--no-default-scuba-dataset` to disable the default data set.
* For backwards compatibility, we still accept `DATASET` starting with `file://` to mean log to a file.
* If any of these options are specified, then the default scuba dataset is ignored.

Reviewed By: clara-9

Differential Revision: D66544989

fbshipit-source-id: 5f13eb9da70865b3772943855c29b7c759d395f1