introduce CachingKeyStore
Summary:
# This diff

Implements the CachingKeyStore which will be the basis for the CachingTreeStore and CachingFileStore implementations. This will allow me to avoid duplicating the caching logic in both of the CachingTree/FileStore implementations

# Context

I'm introducing a new CachingFileStore implementation that will look very similar to the current CachingTreeStore implementation. I'm doing a number of refactors to avoid code duplication. In the end, we'll have:

* CachingKeyStore (generic keystore caching layer)
* CachingFileStore which implements the FileStore traits
* CachingTreeStore which implements the TreeStore traits

The last 2 in the list will wrap a CachingKeyStore and delegate the caching operations to that wrapped store.

Reviewed By: muirdm

Differential Revision: D64670783

fbshipit-source-id: 7c6e175af73a84d9307e53fcfc2d30add18d225e