allow CachingKeyStore to be unbounded
Summary:
# This diff

Allows the CachingKeyStore structure to use a HashMap instead of an LruCache so that the cache can be unbounded for some use cases.

# Context

Some use cases want to add to Caching{File,Tree}Stores without having to worry about items being dropped from the cache. These use cases essentially use the CachingKeyStore as an in-memory (and therefore ephemeral) storage for objects that are not meant to be persisted on disk. These use cases still want objects to persist until the short lived Sapling process ends, and therefore they want to avoid these objects being evicted from cache.

To do this, we can allow the CachingKeyStore to be unbounded if the specified size is 0. In this case, we'll use a HashMap instead of a LRU Cache

Reviewed By: muirdm

Differential Revision: D66271319

fbshipit-source-id: cdad781af7916ea4206c009c701c333a4d0d290b