CachingTreeStore: use CachingKeyStore
Summary:
# This diff

Refactors the CachingTreeStore to use the new CachingKeyStore implementation. The CachingTreeStore simply wraps the CachingKeyStore and delegates most operations to the underlying key store. TreeStore trait methods still get forwarded to the underlying TreeStore

# Context

I'm introducing a new CachingFileStore implementation that will look very similar to the current CachingTreeStore implementation. I'm doing a number of refactors to avoid code duplication. In the end, we'll have:

* CachingKeyStore (generic keystore caching layer)
* CachingFileStore which implements the FileStore traits
* CachingTreeStore which implements the TreeStore traits

The last 2 in the list will wrap a CachingKeyStore and delegate the caching operations to that wrapped store.

Reviewed By: muirdm

Differential Revision: D64670777

fbshipit-source-id: d0ce5a3d3a54e13a7017ced336a676674c88e0e5