introduce CachingFileStore
Summary:
# This diff

Introduces a CachingFileStore which uses the new CachingKeyStore implementation. The CachingKeyStore simply wraps the CachingKeyStore and delegates most operations to the underlying key store.

However, since FileStore operations are inherrently aux aware (aux data can be derived as long as content is provided), we can easily make the FileStore trait methods caching-aware. The next diffs in the stack will do that.

# Context

I'm introducing a new CachingFileStore implementation that will look very similar to the current CachingTreeStore implementation. I'm doing a number of refactors to avoid code duplication. In the end, we'll have:

* CachingKeyStore (generic keystore caching layer)
* CachingFileStore which implements the FileStore traits
* CachingTreeStore which implements the TreeStore traits

The last 2 in the list will wrap a CachingKeyStore and delegate the caching operations to that wrapped store.

Reviewed By: muirdm

Differential Revision: D64670785

fbshipit-source-id: 59445f12d471915e6f9e4a1ccedbef29d24d1a7f