TestStore: implement FileStore trait
Summary:
# This diff

Implements the "FileStore" trait for TestStore so that it can be used later in the stack to test the new CachingFileStore implementation

# Context

I'm introducing a new CachingFileStore implementation that will look very similar to the current CachingTreeStore implementation. I'm doing a number of refactors to avoid code duplication. In the end, we'll have:

* CachingKeyStore (generic keystore caching layer)
* CachingFileStore which implements the FileStore traits
* CachingTreeStore which implements the TreeStore traits

The last 2 in the list will wrap a CachingKeyStore and delegate the caching operations to that wrapped store.

Reviewed By: muirdm

Differential Revision: D64670782

fbshipit-source-id: 47344c43168c559b450ca525e4c5495b6395e722