cas: make non-REMOTE fetches no-ops
Summary: EdenFS typically first fetches contents in local-only mode from the backingstore (for lower latency when data is cached locally), and only if not available locally, shuffles the request off to a separate thread pool that batches things up to fetch remoteily. The backingstore treats the CAS backend as a remote backend, so in CAS mode, the local fetch almost never finds data and is a waste of time. Let's just make it a no-op by default.

Reviewed By: quark-zju

Differential Revision: D66679366

fbshipit-source-id: e052c237b58da3176a8528db6f31b1807578aff7