Fix ServiceError::Poll responses
Summary:
## This stack

Improve logging to help troubleshooting.

## This diff

D64472441 introduced a `PollError` exception  but I never properly tested it end to end, as shown by the remaining `todo!`.

An integration test for this is going to be somewhat unpleasant; I guess I'll have to add a JK in the right place to force a failure.

Reviewed By: RajivTS

Differential Revision: D66546858

fbshipit-source-id: 484ba15ead2b7d24aa50773b57876c49b3aae388