Fix ODS counters for queue stats
Summary:
The previous implementation has a serious bug: it only updates counters for statuses that are present in the results from `get_queue_stats` (i.e. in the DB).

That creates a scenario where a counter (e.g. `async_requests.worker.stats.queue.age_s.new`) can get stuck to a high value; and then never reset to 0. This is particularly apparent because this code is only called infrequently; so it is very prone to sampling bias.

This diff changes that logic. Hardcoding the possible states is not great, but it's reasonable and safe: if we ever add new states and forget to add them, that will be easy to notice i.e. we just won't have those stats at all.

Reviewed By: mitrandir77, clara-9

Differential Revision: D66698389

fbshipit-source-id: 9dc06439b7e2b72fa8fcc80d197a9de843d29a96