Correctly log failures to save
Summary:
## This stack

Improve logging to help troubleshooting.

## This diff

The worker logic handles edge cases where we were able to process the requests correctly; but saving (to either blobstore or DB) failed.

Before this diff we would love the success to Scuba, with no indication of the save failure. That would make debugging hard. After this diff we will log to overall status as a failure, and the details logged to the `error` column.

Note that after this error the request will still be in the queue and eventually be retried.

Reviewed By: mitrandir77

Differential Revision: D66699480

fbshipit-source-id: 89102f75b25a788d4db3a7cb7d34a9156ac88a65