land_service: use DbLocator to find master db region instead of smc fake tiers
Summary: Use of smc fake tiers is discouraged and we have to whitelist our binaries in order to use it. This is error prone as the whitelist applies to paths so any changes in the path of the build rules for our binaries can cause failures. Use `DbLocator` instead to find the region of db master. I'm gating this behind a JK for easy rollback.

Reviewed By: clara-9

Differential Revision: D66710123

fbshipit-source-id: 6858faa12d40dc4d7f516ecfe2687feafe34a0b6