Configure WBC in Mononoke Git server to derive Git derived data types
Summary: Although we derive git derived data types during Git push, this does not apply for commits that are landed through SCS. Although right now this is just the case for cross synced repos like `whatsapp/server`, in the future commits could also get landed through SCS for aosp repos. This means that we need a holistic way to ensure that we do derive the needed commits for any Git request that we could receive.

Reviewed By: lmvasquezg

Differential Revision: D66769325

fbshipit-source-id: 2037e78a2b77d361e9f585ea2b152331c635de08