revisionstore: don't use fbinit::assume_init()
Summary:
Why assume when you can make sure it is initialized?

The answer is that originally it was not actually initialized (and we had an unsafe "unsafe" block), but I fixed it in D60475562.

We still need to handle when fb is not initialized for some tests and other entry points (OSS mode, I think).

Reviewed By: zzl0

Differential Revision: D66731807

fbshipit-source-id: 06837221248da58c6ca2627863e0ab8907cb4949