Back out "cas: make non-REMOTE fetches no-ops"
Summary: This change made operations error out when CAS was enabled and EdenFS had experimental.allow-remote-get-batch=false (as is set in production). In particular, the issue was we perform a LOCAL_ONLY scmstore query followed by a REMOTE_ONLY query, but since LOCAL_ONLY became a no-op, data that was only available locally could not be fetched.

Reviewed By: liubov-dmitrieva

Differential Revision: D66775197

fbshipit-source-id: 38093d720fe77c95ffa8254eeff737bb2058f1dd