Introduce method to generate ref content mapping in gitimport and remote gitimport
Summary: As part of `gitimport` and `remote-gitimport`, we would now need to support storing mappings of git refs pointing to content git objects (i.e. trees and blobs). This diff adds that method and implements it in `gitimport`. Follow up diffs will add support for implementing it through SCS in `remote-gitimport`

Reviewed By: mitrandir77

Differential Revision: D66656227

fbshipit-source-id: eb0d8dd2061e6e835164f538a53e4c7be936bce1