Run sync operations in parallel
Summary:
## This stack

Track and down methods that are holding up the reactor and optimize them.

## This diff

This code is loading both sides of a diff from blobstore concurrently is the right thing. In practice, most of the cost comes from the Thrift deserialization, which is synchronous. Worse, `try_join` does not yield back to Tokio. The combination means blocking the reactor for 150-200ms or even longer.

With this diff we'll actually run them in parallel, saving both max poll time and quite likely wall clock time too.

Reviewed By: markbt

Differential Revision: D66761433

fbshipit-source-id: 5d44506a453697d7a45afe3bd811152ba0e21982