metrics: add basic ods support to Counter
Summary:
Now (when compiled by buck in fbsource) the Counter type also contains an ODS counter that gets updated as the Counter is updated.

Rather than using the dynamic counter in stats::define_stats macro, I directly instantiate a BoxSingletonCounter and store it in the Counter. This avoids computing the String metric name every time when incrementing.

Reviewed By: quark-zju

Differential Revision: D66848953

fbshipit-source-id: 71858669fb3c4e55339ed52bd2a5023aa7927e6b