dirstate: fix the bytes format error
Summary:
When I was running "sl doctor", I got the following error:

```
unexpected error while checking for problems: Traceback (most recent call last):
  File "eden/fs/cli/doctor/check_hg.py", line 405, in check_hg
    if checker.check():
  File "eden/fs/cli/doctor/check_hg.py", line 50, in check
    self.errors = self.check_for_error()
  File "eden/fs/cli/doctor/check_hg.py", line 109, in check_for_error
    self._get_old_dirstate_info(errors)
  File "eden/fs/cli/doctor/check_hg.py", line 127, in _get_old_dirstate_info
    parents, tuples_dict, copymap = eden.dirstate.read(f, str(self.path))
  File "eden/dirstate.py", line 190, in read
    "Unexpected header byte " "when reading {:s}: 0x{:x}.".format(
TypeError: unsupported format string passed to bytes.__format__
```

The issue is that it is trying to format a `bytes` object (`header`) using a
string format that expects an integer. We need to convert the `header` object
to an integer first.

Reviewed By: MichaelCuevas

Differential Revision: D66880923

fbshipit-source-id: 6867e4eaaf362886ac3c7e6ad1a2314e1020124f