add x2p field to payload
Summary:
# This diff

Adds a "x2p" field to the clicat payload that specifies whether x2p should be used or not.

# Context

Currently, calls to the XInternGraphCommitCloudPollWorkspaceUpdateInternGraphJsonController on interngraph are functioning very unreliably. It's an X2P request to interngraph that isn't marked as such or supported.

It's crucial that we fix this so we can unblock another team fixing a X2P vulnerability.

Differential Revision: D66901312

fbshipit-source-id: 204e0e59be334cfc66695ec111c74c0bdf1c4764