Fix hg recover throwing an error
Summary:
# Bug Description
When there is an interrupted transaction, a journal file will be left in the .hg/store file
We use hg recover to fix this. When hg recover is run with no interrupted transactions, it will return with error code 1.
We check the existence of the journal file to determine if we should run hg recover inside eden doctor. If something cleans up the journal file before we do, an about hg recover returning error code 1 is reported.

# This Diff
Catch the error when running hg recover. If it is the one where there is nothing to fix, ignore the error

Reviewed By: genevievehelsel

Differential Revision: D66900023

fbshipit-source-id: 5a0c4c808ae8447e84cbd37f055c9ea0f3593ac6