Update network doctor to show speed results and explain timeout
Summary:
# Context
Users with a slow network can hit the timeout on network doctor. This currently can be somewhat cryptic as to why it's timing out
# This Diff
Explains what the timout is, and prints out the network doctor stdout/stderr

# Technical Details
tell user what NETWORK_TIMEOUT is. Explain that hitting this timeout may be due to a slow network. Print out stdout from hg debugnetwork commands so user can see if this is the case

Reviewed By: genevievehelsel

Differential Revision: D66771101

fbshipit-source-id: 30742c51f26b9841c4a9cb06b7fe343d3ed39270