identity: add `resolve_full_dot_dir` to handle complex cases
Summary:
In complex cases, such as dotgit submodule, it's not as simple as
`root.join(dot_dir)`. The `dot_dir` is `.git/sl`, the `root` has a `.git` file
with content like `gitdir: foo/bar`. Add a function in `identity` to ease the
callsite needing the information.

This diff only adds the API. It is not yet used in other places yet.

Reviewed By: sggutier

Differential Revision: D66911301

fbshipit-source-id: 2fcd9be7d8d32e9db4c19297de1ff7e1a4a72eab