remove side panel gating
Summary:
Now that we've rolled the V1 out and everyone with this feature should see the panel (panel is an essential part of the UX now and I don't think we'll roll it back). I'm removing the 'devFeature' gating for it. I'm keeping the enableDevFeature flag in code so we can later reuse to gate any non-ready features.

Also, refactoring the conditional instanitiation to outside of the the constructor while I'm at it. I think it's cleaner this way.

Differential Revision: D66968009

fbshipit-source-id: f0c4acd6a1f920d3fb96c7194c42940cace3e28f