refactored fetch comment and display multi reigon code change in inset
Summary:
Refactored fetch comment so that we create "patch" object for each of the code change section. So we can later feed it into the DiffView

There is a minor bug in this diff where in the Diff View we only show the lines that changes and filtered out all the lines in between that arent changed. Will implement that in a follow up diff.

Differential Revision: D66957463

fbshipit-source-id: b75c6319e6b44c255988b7f7fbe5e26c842c034d