teach FileAccessLogger how to log directory accesses
Summary:
This diff teaches the `FileAccessLogger` how to log directory accesses. The changes include adding a `dtype` field to the `FileAccess` struct and modifying the `logFileAccess` method to log directory accesses based on the `dtype` field. The changes are made in the `FileAccessLogger.cpp` and `InodeBase.cpp` files.

This change is necessary because the `FileAccessLogger` previously assumed only `FileInode`s would be logged, but we now must handle `TreeInodes` to support Windows usage of predictive prefetching

Reviewed By: MichaelCuevas

Differential Revision: D66778516

fbshipit-source-id: 882953b1d2c53373bb94b89a07198988f32b33f7