Optimize commit_file_diffs by running xdiff::diff_unified in a different thread
Summary:
## This stack

Track and down methods that are holding up the reactor and optimize them.

## This diff

`xdiff::diff_unified` is synchronous and does a lot of work to format a diff. While we should optimize it, it is older code so that will realistically take a while. Running it on a separate, non-executor thread gets the job done.

Reviewed By: lmvasquezg

Differential Revision: D67025628

fbshipit-source-id: 1f946e8f7aabc0eb8f632a5724ed3a608ec6d877