merge: fix path-remapping for merge.ACTION_CREATED
Summary:
When doing merge with `sl merge --from-path foo --to-path bar -r A`, the diff operation produces paths like "bar/file", but we need to recover the original path "foo/file" to look up the file context in A.

The ACTION_CREATED action did not support a different "local" and "remote" filename. Extend their args to include the remote file name (which differs when merging between directories). This diff is similar to D60027064, which extended the args of actions like ACTION_CREATED_MERGE.

Besides that, this diff also fix a bug in checkpathconflicts(). The ACTION_CREATED_MERGE's `args[0]` was not flags anymore after D60027064.

Reviewed By: quark-zju

Differential Revision: D66986091

fbshipit-source-id: 0c983e294dc6899bcc61ec3465d8dbd041112114