Send scmstore Sapling counters through OBC API
Summary:
## Background:
 Currently we are using FB303 API to send stats to ODS. We are migrating a few stats to OBC API, because it makes it possible for us to collect some ODS stats from sandcastle. This diff D65554643 describe the pros and cons of OBC API over FB303 API, and how we can have same performance by migrating to OBC API.
The `scmstore` counters are collected in Sapling. However, we are reporting the counters on Eden. Moving the `scmstore` counters to OBC API not only enable them on sandcastle, but also gives us a histogram of `.sum.60` instead of accumulated values.
The next diff send the `scmstore` counters through OBC API. In this diff, I want to add a new Sapling config. Then we can control roll out of this new counter types

## This diff:
Update `ods::increment()` function to bump OBC API counters in addition to incrementing the ODS counters. Therefore, we can have the histogram value in 60 sec in addition to the accumulated values. It bumps ODS directly from Sapling, and don't transfer counters to eden for collecting ODS.

## Note:
The changes are behind a config which is false now, then it doesn't have any performance regression. I will open it gradually. Also, bumping ODS doesn't break any thing.

Reviewed By: quark-zju

Differential Revision: D66969706

fbshipit-source-id: 1a8ee52f00e330f85718fd10ea4b9025cbb74305