Add test for folder rename - Windows
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff
This diff adds tests for folder renaming on windows. This is allowed for uncommited folders but not for commited ones

# Technical Details
Update setup to create a folder and a subfile
Add tests for:
rename folder
rename folder with file
rename folder with file- commited

# Discussion Points
Currently there is no file removed notification for the file in the old folder. Should there be?

Reviewed By: MichaelCuevas

Differential Revision: D66994320

fbshipit-source-id: 3915fe17bcb6259d1f7539ad89ad7696fad53879