Directory modified tests: chown/chmod
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff
This diff adds tests for folder modifications caused by chown/chmod

# Technical Details
New files:
chown requires sudo to change to the nobody:nobody user. This means that the test does not work on Sandcastle since this user does not exist on it. Hence it is disabled on all platforms

# Discussion Points

Reviewed By: jdelliot

Differential Revision: D66984085

fbshipit-source-id: 183f4a4fe4ad4b8653bf0d7f3d022ffa0e8fe2c9