UncommittedChanges: add "Absorb" button
Summary:
Add a button to perform the absorb operation. It does not do anything yet.

I thought about putting the button in the "stack actions" row. For absorb it
seems the user would be more focused on the "pending changes" selection, less
about the stack of commits. So I put the button at the UncommittedChanges section.

The code does seem to make sense to live in StackActions, since it will likely
share some states.

Reviewed By: zzl0

Differential Revision: D67045935

fbshipit-source-id: 12a79f0af12b6fceeefa9402153ce9ddeae1d8ff