modern sync: upload hg changeset
Summary:
This is not a final version, this is simply to make sure all elements of the filenodes are conserved through edenapi. This is missing two thing:
1. Correct Parents expression (mentioned by Mark in D67034907)
2. Stream/ pipe through channel so we don't keep everything in memory

But hey, it's a god POC, putting up for a review in case something is terribly wrong, otherwise I can address 1 and 2 in separate diff (will decide after chat about the parents issue)

This also adds a condition so we don't try to add hg mutations if none are sent.

With this we can now make sure all relevant data has been sent to the server and we can start optimising :)

Reviewed By: clara-9

Differential Revision: D67123774

fbshipit-source-id: dd042bdb45d49e5c4987a974a6d81c75042c2633