Refactor code to be easier to optimize
Summary:
## This stack

We noticed that the poll time is pretty high on some async methods, and that is affecting other calls as well. Let's try to optimize that.

## This diff

The next diff in the stack is going to be tricky, because of some async/sync wizardry. Do this simple refactoring first to keep things simple to review.

Reviewed By: singhsrb

Differential Revision: D67191474

fbshipit-source-id: 3119de0a25cd19d5deaca78beda447c9786cb701