Run expensive code on a separate thread
Summary:
## This stack

We noticed that the poll time is pretty high on some async methods, and that is affecting other calls as well. Let's try to optimize that.

## This diff

The bulk of the cost of `commit_sparse_profile_delta` comes from a complex section of synchronous code; tuning that would be a fair amount of work, but we can at least ensure we don't block an executor.

Reviewed By: singhsrb

Differential Revision: D67136194

fbshipit-source-id: d6c39ba841c34f4c6c323c62a4c9ef4c859236ef