Remove overlay file/dir counters
Summary:
These counters are not working as expected. The values are mostly negative https://fburl.com/canvas/8hgwlts5
The root cause is that `dirCount` and `fileCount` are not ++/-- in the right places. We need to find a better way probably from `TreeInode` or `FileInode` to count the total number of file/dir which are loaded in OverLay.
For now, we should clean up these counters. As these counters are per repo, they can spam our eden ODS category with lots of trash values in time-series.
Later, I will find a better way for counting the items in Overlay, and publish these counters again.

Reviewed By: genevievehelsel

Differential Revision: D67213689

fbshipit-source-id: 66a6566101ead446870413ca8eb8b3fa3717519e