ChangesV2 test for truncated journal
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff
This diff adds tests around truncated journals

If the journal becomes truncated, all changes before that are lost.
When passing in a position before the truncation, only the truncation is returned
When passing in a position after the truncation, you get the expected results

# Technical Details
Tests that
- Truncated journal appears when the journal is truncated
- Test that changes before the truncation are lost
- Test that changes after the truncation are lost when passing in a position before the truncation
- Test that changes after the truncation are present when passing in a position after the truncation

# Discussion Points

Reviewed By: jdelliot

Differential Revision: D67221851

fbshipit-source-id: ad0fcad738bac4a7636c1bffe8603f3bb463df84