ChangesV2 test for commit transition
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff
This diff adds tests around commit transitions

A commit transition occurs when the user changes commits.
The format of a commit transition is from=the previous commit, new=the current commit
This is because we walk backwards through the journal when reading it.

# Technical Details
Tests that
- Commit transition shows up when creating commits
- Commit transition shows up when changing commits

# Discussion Points

Reviewed By: jdelliot

Differential Revision: D67220177

fbshipit-source-id: 32786fd03c0e3b75900dc2c434e0c1c5094d060d